Fix mission ID type in JdbcMoonMissionRepository; update label in Main class mission details output

+++ src/main/java/com/example/JdbcMoonMissionRepository.java
@@ -52,7 +52,7 @@ public class JdbcMoonMissionRepository implements MoonMissionRepository{
                 if(rs.next()){
                     //Map to MoonMission object:
                     MoonMission mission= new MoonMission(
-                            rs.getInt("mission_id"),
+                            rs.getLong("mission_id"),
                             rs.getString("spacecraft"),
                             rs.getObject("launch_date", LocalDate.class),
                             rs.getString("carrier_rocket"),
